requirements.txt
cx-Oracle==8.3.0
openai==1.12.0
python-dotenv==1.0.1 